[FEAT][#34] 유저 프로필 View 재사용을 위해 공통View로 생성
