+++ PersonalInformation/PersonalInformationController.swift
@@ -18,78 +18,13 @@ class PersonalInformationController: UIViewController {
                         2: ("결석", 0, -30)]
     // indexPath.row: (title.text, result.text, score?.text)
     
-    private let username = "김넥터"
-    private let userJob = "designer"
-    private let userYear = 22
-    private let email = "nexters@naver.com"
-    
+    private let profileView = ProfileView()
     private let historyTableView = UITableView()
     private let resultCollectionContainerView = UIView()
     private let resultCollectionView = UICollectionView(frame: .zero, collectionViewLayout: UICollectionViewFlowLayout.init())
     
     private lazy var hintLatePenalty = setupHintPenalty("-5점")
     private lazy var hintAbsentPenalty = setupHintPenalty("통보 -10 / 무단 -15")
-    
-    private lazy var profileImageView: UIImageView = {
-        let imageView = UIImageView(frame: CGRect(x: 0, y: 0, width: 93, height: 93))
-        imageView.image = UIImage(named: "\(userJob)")!
-        imageView.contentMode = .scaleAspectFill
-        imageView.layer.cornerRadius = 93 / 2
-        imageView.clipsToBounds = true
-        imageView.layer.borderWidth = 1
-        imageView.layer.borderColor = UIColor.rgba(208, 208, 208, 1).cgColor
-        
-        return imageView
-    }()
-    
-    private lazy var userActiveStatus: UIView = {
-        let statusView = UIView()
-        
-        let status = UILabel()
-        status.text = "\(userYear)기 참여중"
-        status.font = .systemFont(ofSize: 13, weight: .bold)
-        status.textColor = .white
-        
-        statusView.addSubview(status)
-        status.snp.makeConstraints { make in
-            make.edges.equalTo(statusView).inset(9)
-        }
-        statusView.backgroundColor = UIColor.rgba(116, 207, 130, 1)
-        statusView.layer.cornerRadius = 5
-        return statusView
-    }()
-    
-    private lazy var userBasicInfomationView: UIStackView = {
-        let usernameLabel = UILabel()
-        let jobLabel = UILabel()
-        let jobLabelView = UIView()
-        let view = UIStackView(arrangedSubviews: [usernameLabel, jobLabelView])
-        view.axis = .horizontal
-        view.spacing = 10
-        
-        usernameLabel.text = username
-        usernameLabel.font = .systemFont(ofSize: 22, weight: .medium)
-        usernameLabel.textColor = .black
-        
-        jobLabel.text = userJob.uppercased()
-        jobLabel.font = .systemFont(ofSize: 13)
-        jobLabelView.addSubview(jobLabel)
-        jobLabel.snp.makeConstraints { make in
-            make.verticalEdges.equalTo(jobLabelView).inset(5)
-            make.horizontalEdges.equalTo(jobLabelView).inset(10)
-        }
-        jobLabelView.backgroundColor = UIColor.rgba(243, 243, 243, 1)
-        
-        return view
-    }()
-    
-    private lazy var userAdditionalInformationView: UILabel = {
-        let label = UILabel()
-        label.text = email
-        label.font = .systemFont(ofSize: 16)
-        label.textColor = UIColor.rgba(118, 118, 118, 1)
-        return label
-    }()
     
     private lazy var tableViewHeader: UIView = {
         let header = UIView(frame: CGRect(x: 0, y: 0, width: view.frame.size.width, height: 38))
@@ -154,10 +89,6 @@ class PersonalInformationController: UIViewController {
         logout.font = .systemFont(ofSize: 15)
         navigationItem.rightBarButtonItem = UIBarButtonItem(customView: logout)
     }
-
-    private func setupUserProfileImage() {
-        view.addSubviews(profileImageView, userActiveStatus, userBasicInfomationView, userAdditionalInformationView)
-    }
     
     private func setupHintPenalty(_ hint: String) -> UILabel {
         let hintLabel = UILabel()
@@ -195,7 +126,7 @@ class PersonalInformationController: UIViewController {
     }
     
     private func setupViews() {
-        setupUserProfileImage()
+        view.addSubview(profileView)
         setupResultCollectionView()
         setupHistoryTableView()
     }
@@ -203,27 +134,15 @@ class PersonalInformationController: UIViewController {
     private func setupLayout() {
         setupViews()
         
-        profileImageView.snp.makeConstraints { make in
-            make.top.equalTo(view.safeAreaLayoutGuide).offset(36)
+        profileView.snp.makeConstraints { make in
+            make.top.equalTo(view.safeAreaLayoutGuide).offset(10)
             make.centerX.equalToSuperview()
-        }
-        userActiveStatus.snp.makeConstraints { make in
-            make.top.equalTo(profileImageView.snp.centerY).offset(10)
-            make.left.equalTo(profileImageView.snp.centerX).offset(10)
-        }
-        
-        userBasicInfomationView.snp.makeConstraints { make in
-            make.top.equalTo(profileImageView.snp.bottom).offset(20)
-            make.centerX.equalToSuperview()
-        }
-        userAdditionalInformationView.snp.makeConstraints { make in
-            make.top.equalTo(userBasicInfomationView.snp.bottom).offset(10)
-            make.centerX.equalToSuperview()
+            make.height.equalTo(200)
         }
         
         resultCollectionContainerView.snp.makeConstraints { make in
             make.horizontalEdges.equalToSuperview()
-            make.top.equalTo(userAdditionalInformationView.snp.bottom).offset(27)
+            make.top.equalTo(profileView.snp.bottom).offset(20)
             make.height.equalTo(135)
         }
