+++ momoIOS/Management/MoimManageController.swift
@@ -1,0 +1,8 @@
+//
+//  MoimManageController.swift
+//  momoIOS
+//
+//  Created by 문다 on 2023/02/09.
+//
+
+import Foundation
